크라우드 펀딩 상세 조회 API 연결

+++ design-system/src/main/java/com/danbam/design_system/component/ProgressBar.kt
@@ -32,10 +32,10 @@ sealed class MakeFundingProgress(val progress: Int) {
 @Composable
 fun IndiStrawProgress(
     modifier: Modifier = Modifier,
-    currentProgress: Float,
+    currentProgress: Long,
     enableText: Boolean = true
 ) {
-    val progressAnimation by animateFloatAsState(targetValue = currentProgress / 100)
+    val progressAnimation by animateFloatAsState(targetValue = currentProgress / 100F)
     Row(
         modifier = modifier,
         verticalAlignment = Alignment.Bottom
